Remove leftover usages of the fake function
The fake function provides no use beyond plain js string interpolation,
which should be used instead. Besides that any function within this
library that would use the fake method as intended would need to import
every module and therefore renders tree shaking useless.

The removed usages did not need the fake function. The `city` function
was merely delegating the format string to `fake` so a user of the
library could just use that directly (or rather string interpolation as
explained).